Split off Glyph API and support subcoordinates
