+++ bokehjs/test/integration/annotations/legend.ts
@@ -3,7 +3,7 @@ import {display, fig} from "../_util"
 import {Legend, LegendItem, LinearAxis} from "@bokehjs/models"
 import {Random} from "@bokehjs/core/util/random"
 import {range} from "@bokehjs/core/util/array"
-import {CircleArgs, LineArgs} from "@bokehjs/api/plotting"
+import {CircleArgs, LineArgs} from "@bokehjs/api/glyph_api"
 import {Orientation} from "@bokehjs/core/enums"
 
 describe("Legend annotation", () => {
