+++ bokehjs/test/integration/regressions.ts
@@ -35,7 +35,8 @@ import {Random} from "@bokehjs/core/util/random"
 import {Matrix} from "@bokehjs/core/util/matrix"
 import {defer} from "@bokehjs/core/util/defer"
 import {encode_rgba} from "@bokehjs/core/util/color"
-import {Figure, MarkerArgs, show} from "@bokehjs/api/plotting"
+import {Figure, show} from "@bokehjs/api/plotting"
+import {MarkerArgs} from "@bokehjs/api/glyph_api"
 import {Spectral11, turbo} from "@bokehjs/api/palettes"
 import {div, offset} from "@bokehjs/core/dom"
